Revert "feat:  Table rendering support for databars (#1212)"
This reverts commit a17cc0eb2b4e8ba9240c891a15b9d4b7659fb721.
