+++ packages/grid/src/GridModel.ts
@@ -4,7 +4,6 @@ import { ModelIndex } from './GridMetrics';
 import { GridColor, GridTheme, NullableGridColor } from './GridTheme';
 import memoizeClear from './memoizeClear';
 import GridUtils, { Token } from './GridUtils';
-import { CellRenderType } from './CellRenderer';
 
 const LINK_TRUNCATION_LENGTH = 5000;
 
@@ -227,10 +226,6 @@ abstract class GridModel<
       return GridUtils.findTokensWithProtocolInText(contentToCheckForLinks);
     }
   );
-
-  renderTypeForCell(column: ModelIndex, row: ModelIndex): CellRenderType {
-    return 'text';
-  }
 }
 
 export default GridModel;
