+++ packages/grid/src/GridRenderer.test.tsx
@@ -3,7 +3,6 @@ import GridModel from './GridModel';
 import GridRenderer from './GridRenderer';
 import MockGridModel from './MockGridModel';
 import GridTheme from './GridTheme';
-import TextCellRenderer from './TextCellRenderer';
 import { LinkToken } from './GridUtils';
 import { GridRenderState } from './GridRendererTypes';
 
@@ -117,10 +116,7 @@ describe('getTokenBoxesForVisibleCell', () => {
       }),
     });
 
-    const textCellRenderer = renderer.getCellRenderer(
-      'text'
-    ) as TextCellRenderer;
-    textCellRenderer.getCachedTruncatedString = jest.fn(
+    renderer.getCachedTruncatedString = jest.fn(
       (
         context: CanvasRenderingContext2D,
         text: string,
@@ -132,14 +128,7 @@ describe('getTokenBoxesForVisibleCell', () => {
   });
 
   it('should return tokens that are visible in the cell', () => {
-    const textCellRenderer = renderer.getCellRenderer(
-      'text'
-    ) as TextCellRenderer;
-    const tokens = textCellRenderer.getTokenBoxesForVisibleCell(
-      0,
-      0,
-      renderState
-    );
+    const tokens = renderer.getTokenBoxesForVisibleCell(0, 0, renderState);
 
     const expectedValue: LinkToken = {
       type: 'url',
@@ -155,14 +144,7 @@ describe('getTokenBoxesForVisibleCell', () => {
   });
 
   it('should return multiple tokens', () => {
-    const textCellRenderer = renderer.getCellRenderer(
-      'text'
-    ) as TextCellRenderer;
-    const tokens = textCellRenderer.getTokenBoxesForVisibleCell(
-      0,
-      2,
-      renderState
-    );
+    const tokens = renderer.getTokenBoxesForVisibleCell(0, 2, renderState);
 
     const expectedValue: LinkToken[] = [
       {
@@ -191,14 +173,7 @@ describe('getTokenBoxesForVisibleCell', () => {
   });
 
   it('should return empty array if there are no tokens', () => {
-    const textCellRenderer = renderer.getCellRenderer(
-      'text'
-    ) as TextCellRenderer;
-    const tokens = textCellRenderer.getTokenBoxesForVisibleCell(
-      0,
-      1,
-      renderState
-    );
+    const tokens = renderer.getTokenBoxesForVisibleCell(0, 1, renderState);
 
     expect(tokens).toHaveLength(0);
   });
@@ -206,14 +181,7 @@ describe('getTokenBoxesForVisibleCell', () => {
   it('should return empty array if context or metrics is null', () => {
     // @ts-expect-error metrics and context usually can't be null
     renderState = makeMockGridRenderState({ metrics: null, context: null });
-    const textCellRenderer = renderer.getCellRenderer(
-      'text'
-    ) as TextCellRenderer;
-    const tokens = textCellRenderer.getTokenBoxesForVisibleCell(
-      0,
-      0,
-      renderState
-    );
+    const tokens = renderer.getTokenBoxesForVisibleCell(0, 0, renderState);
 
     expect(tokens).toHaveLength(0);
   });
