+++ packages/grid/src/GridRenderer.ts
@@ -1,18 +1,15 @@
 import clamp from 'lodash.clamp';
-import { ColorUtils, getOrThrow } from '@deephaven/utils';
+import { ColorUtils, EMPTY_ARRAY, getOrThrow } from '@deephaven/utils';
 import memoizeClear from './memoizeClear';
-import GridUtils from './GridUtils';
+import GridUtils, { Token, TokenBox } from './GridUtils';
 import GridColorUtils from './GridColorUtils';
 import { isExpandableGridModel } from './ExpandableGridModel';
 import { GridColor, GridColorWay, NullableGridColor } from './GridTheme';
-import { Coordinate, VisibleIndex } from './GridMetrics';
+import { BoxCoordinates, Coordinate, VisibleIndex } from './GridMetrics';
 import { isEditableGridModel } from './EditableGridModel';
 import GridColumnSeparatorMouseHandler from './mouse-handlers/GridColumnSeparatorMouseHandler';
 import { BoundedAxisRange } from './GridAxisRange';
-import { DEFAULT_FONT_WIDTH, GridRenderState } from './GridRendererTypes';
-import CellRenderer, { CellRenderType } from './CellRenderer';
-import DataBarCellRenderer from './DataBarCellRenderer';
-import TextCellRenderer from './TextCellRenderer';
+import { GridRenderState } from './GridRendererTypes';
 
 type NoneNullColumnRange = { startColumn: number; endColumn: number };
 
@@ -27,15 +24,14 @@ type NoneNullRowRange = { startRow: number; endRow: number };
  * your own methods to customize drawing of the grid (eg. Draw icons or special features)
  */
 export class GridRenderer {
+  // Default font width in pixels if it cannot be retrieved from the context
+  static DEFAULT_FONT_WIDTH = 10;
+
   // Default radius in pixels for corners for some elements (like the active cell)
   static DEFAULT_EDGE_RADIUS = 2;
 
   // Default width in pixels for the border of the active cell
   static ACTIVE_CELL_BORDER_WIDTH = 2;
-
-  protected textCellRenderer = new TextCellRenderer();
-
-  protected dataBarCellRenderer = new DataBarCellRenderer();
 
   /**
    * Truncate a string to the specified length and add ellipses if necessary
@@ -122,7 +118,7 @@ export class GridRenderer {
     context: CanvasRenderingContext2D,
     str: string,
     width: number,
-    fontWidth = DEFAULT_FONT_WIDTH,
+    fontWidth = GridRenderer.DEFAULT_FONT_WIDTH,
     truncationChar?: string
   ): string {
     if (width <= 0 || str.length <= 0) {
@@ -997,28 +993,242 @@ export class GridRenderer {
     context.restore();
   }
 
+  /**
+   * Gets textWidth and X-Y position for a specific cell
+   * The textWidth returned is the width that the text can occupy accounting for any other cell markings
+   * The width accounts for tree table indents and cell padding, so it is the width the text may consume
+   *
+   * @param state GridRenderState to get the text metrics for
+   * @param column Column of cell to get text metrics for
+   * @param row Row of cell to get text metrics for
+   * @returns Object with width, x, and y of the text
+   */
+  getTextRenderMetrics(
+    state: GridRenderState,
+    column: VisibleIndex,
+    row: VisibleIndex
+  ): {
+    width: number;
+    x: number;
+    y: number;
+  } {
+    const { metrics, model, theme } = state;
+    const {
+      firstColumn,
+      allColumnXs,
+      allColumnWidths,
+      allRowYs,
+      allRowHeights,
+      modelRows,
+      modelColumns,
+    } = metrics;
+    const {
+      cellHorizontalPadding,
+      treeDepthIndent,
+      treeHorizontalPadding,
+    } = theme;
+
+    const modelRow = getOrThrow(modelRows, row);
+    const modelColumn = getOrThrow(modelColumns, column);
+    const textAlign = model.textAlignForCell(modelColumn, modelRow);
+    const x = getOrThrow(allColumnXs, column);
+    const y = getOrThrow(allRowYs, row);
+    const columnWidth = getOrThrow(allColumnWidths, column);
+    const rowHeight = getOrThrow(allRowHeights, row);
+    const isFirstColumn = column === firstColumn;
+    let treeIndent = 0;
+    if (
+      isExpandableGridModel(model) &&
+      model.hasExpandableRows &&
+      isFirstColumn
+    ) {
+      treeIndent =
+        treeDepthIndent * (model.depthForRow(row) + 1) + treeHorizontalPadding;
+    }
+    const textWidth = columnWidth - treeIndent;
+    let textX = x + cellHorizontalPadding;
+    const textY = y + rowHeight * 0.5;
+    if (textAlign === 'right') {
+      textX = x + textWidth - cellHorizontalPadding;
+    } else if (textAlign === 'center') {
+      textX = x + textWidth * 0.5;
+    }
+    textX += treeIndent;
+
+    return {
+      width: textWidth - cellHorizontalPadding * 2,
+      x: textX,
+      y: textY,
+    };
+  }
+
   drawCellContent(
     context: CanvasRenderingContext2D,
     state: GridRenderState,
     column: VisibleIndex,
-    row: VisibleIndex
+    row: VisibleIndex,
+    textOverride?: string
   ): void {
-    const { metrics, model } = state;
-    const { modelColumns, modelRows } = metrics;
+    const { metrics, model, theme } = state;
+    const {
+      firstColumn,
+      fontWidths,
+      modelColumns,
+      modelRows,
+      allRowHeights,
+    } = metrics;
+    const { textColor } = theme;
+    const rowHeight = getOrThrow(allRowHeights, row);
     const modelRow = getOrThrow(modelRows, row);
     const modelColumn = getOrThrow(modelColumns, column);
-    const renderType = model.renderTypeForCell(modelColumn, modelRow);
-    const cellRenderer = this.getCellRenderer(renderType);
-    cellRenderer.drawCellContent(context, state, column, row);
+    const text = textOverride ?? model.textForCell(modelColumn, modelRow);
+    const truncationChar = model.truncationCharForCell(modelColumn, modelRow);
+    const isFirstColumn = column === firstColumn;
+
+    if (text && rowHeight > 0) {
+      const textAlign = model.textAlignForCell(modelColumn, modelRow) || 'left';
+      context.textAlign = textAlign;
+
+      const color =
+        model.colorForCell(modelColumn, modelRow, theme) || textColor;
+      context.fillStyle = color;
+
+      context.save();
+
+      const {
+        width: textWidth,
+        x: textX,
+        y: textY,
+      } = this.getTextRenderMetrics(state, column, row);
+
+      const fontWidth =
+        fontWidths.get(context.font) ?? GridRenderer.DEFAULT_FONT_WIDTH;
+      const truncatedText = this.getCachedTruncatedString(
+        context,
+        text,
+        textWidth,
+        fontWidth,
+        truncationChar
+      );
+
+      const tokens = model.tokensForCell(
+        modelColumn,
+        modelRow,
+        truncatedText.length
+      );
+
+      if (truncatedText) {
+        let tokenIndex = 0;
+        let textStart = 0;
+        let left = textX;
+        const { actualBoundingBoxDescent } = context.measureText(truncatedText);
+
+        while (textStart < truncatedText.length) {
+          const nextToken = tokens[tokenIndex];
+          const token = textStart === nextToken?.start ? nextToken : null;
+          const textEnd =
+            token?.end ?? nextToken?.start ?? truncatedText.length;
+          const value = truncatedText.substring(textStart, textEnd);
+          const { width } = context.measureText(value);
+          const widthOfUnderline = value.endsWith('…')
+            ? context.measureText(value.substring(0, value.length - 1)).width
+            : width;
+
+          // Set the styling based on the token, then draw the text
+          if (token != null) {
+            context.fillStyle = theme.hyperlinkColor;
+            context.fillText(value, left, textY);
+            context.fillRect(
+              left,
+              textY + actualBoundingBoxDescent,
+              widthOfUnderline,
+              1
+            );
+          } else {
+            context.fillStyle = color;
+            context.fillText(value, left, textY);
+          }
+
+          left += width;
+          textStart = textEnd;
+          if (token != null) tokenIndex += 1;
+        }
+      }
+      context.restore();
+    }
+
+    if (
+      isFirstColumn &&
+      isExpandableGridModel(model) &&
+      model.hasExpandableRows
+    ) {
+      this.drawCellRowTreeMarker(context, state, row);
+    }
   }
 
-  getCellRenderer(renderType: CellRenderType): CellRenderer {
-    switch (renderType) {
-      case 'dataBar':
-        return this.dataBarCellRenderer;
-      default:
-        return this.textCellRenderer;
+  drawCellRowTreeMarker(
+    context: CanvasRenderingContext2D,
+    state: GridRenderState,
+    row: VisibleIndex
+  ): void {
+    const { metrics, model, mouseX, mouseY, theme } = state;
+    const {
+      firstColumn,
+      gridX,
+      gridY,
+      allColumnXs,
+      allColumnWidths,
+      allRowYs,
+      allRowHeights,
+      visibleRowTreeBoxes,
+    } = metrics;
+    const { treeMarkerColor, treeMarkerHoverColor } = theme;
+    const columnX = getOrThrow(allColumnXs, firstColumn);
+    const columnWidth = getOrThrow(allColumnWidths, firstColumn);
+    const rowY = getOrThrow(allRowYs, row);
+    const rowHeight = getOrThrow(allRowHeights, row);
+    if (!isExpandableGridModel(model) || !model.isRowExpandable(row)) {
+      return;
     }
+
+    const treeBox = getOrThrow(visibleRowTreeBoxes, row);
+    const color =
+      mouseX != null &&
+      mouseY != null &&
+      mouseX >= gridX + columnX &&
+      mouseX <= gridX + columnX + columnWidth &&
+      mouseY >= gridY + rowY &&
+      mouseY <= gridY + rowY + rowHeight
+        ? treeMarkerHoverColor
+        : treeMarkerColor;
+
+    this.drawTreeMarker(
+      context,
+      state,
+      columnX,
+      rowY,
+      treeBox,
+      color,
+      model.isRowExpanded(row)
+    );
+  }
+
+  drawTreeMarker(
+    context: CanvasRenderingContext2D,
+    state: GridRenderState,
+    columnX: Coordinate,
+    rowY: Coordinate,
+    treeBox: BoxCoordinates,
+    color: GridColor,
+    isExpanded: boolean
+  ): void {
+    const { x1, y1, x2, y2 } = treeBox;
+    const markerText = isExpanded ? '⊟' : '⊞';
+    const textX = columnX + (x1 + x2) * 0.5 + 0.5;
+    const textY = rowY + (y1 + y2) * 0.5 + 0.5;
+    context.fillStyle = color;
+    context.textAlign = 'center';
+    context.fillText(markerText, textX, textY);
   }
 
   drawCellRowTreeDepthLines(
@@ -1084,6 +1294,24 @@ export class GridRenderer {
       context.stroke();
     }
   }
+
+  getCachedTruncatedString = memoizeClear(
+    (
+      context: CanvasRenderingContext2D,
+      text: string,
+      width: number,
+      fontWidth: number,
+      truncationChar?: string
+    ): string =>
+      GridRenderer.truncateToWidth(
+        context,
+        text,
+        width,
+        fontWidth,
+        truncationChar
+      ),
+    { max: 10000 }
+  );
 
   getCachedBackgroundColors = memoizeClear(
     (backgroundColors: GridColorWay, maxDepth: number): GridColor[][] =>
@@ -1537,7 +1765,8 @@ export class GridRenderer {
       white,
     } = theme;
     const { fontWidths, width } = metrics;
-    const fontWidth = fontWidths.get(context.font) ?? DEFAULT_FONT_WIDTH;
+    const fontWidth =
+      fontWidths.get(context.font) ?? GridRenderer.DEFAULT_FONT_WIDTH;
 
     const maxWidth = columnWidth - headerHorizontalPadding * 2;
     const maxLength = maxWidth / fontWidth;
@@ -2751,6 +2980,153 @@ export class GridRenderer {
 
     context.translate(-barLeft, -barTop);
   }
+
+  /**
+   * Gets the token boxes that are visible in the cell
+   * @param column The visible column
+   * @param row The visible row
+   * @param state The GridRenderState
+   * @returns An array of TokenBox of visible tokens or empty array with coordinates relative to gridX and gridY
+   */
+  getTokenBoxesForVisibleCell(
+    column: VisibleIndex,
+    row: VisibleIndex,
+    state: GridRenderState
+  ): TokenBox[] {
+    const { metrics, context, model, theme } = state;
+
+    if (context == null || metrics == null) {
+      return (EMPTY_ARRAY as unknown) as TokenBox[];
+    }
+
+    const { modelRows, modelColumns } = metrics;
+    const modelRow = getOrThrow(modelRows, row);
+    const modelColumn = getOrThrow(modelColumns, column);
+
+    const text = model.textForCell(modelColumn, modelRow);
+    const { width: textWidth, x: textX, y: textY } = this.getTextRenderMetrics(
+      state,
+      column,
+      row
+    );
+
+    const { fontWidths } = metrics;
+
+    // Set the font and baseline and change it back after
+    context.save();
+    this.configureContext(context, state);
+
+    const fontWidth =
+      fontWidths?.get(context.font) ?? GridRenderer.DEFAULT_FONT_WIDTH;
+    const truncationChar = model.truncationCharForCell(modelColumn, modelRow);
+    const truncatedText = this.getCachedTruncatedString(
+      context,
+      text,
+      textWidth,
+      fontWidth,
+      truncationChar
+    );
+
+    const {
+      actualBoundingBoxAscent,
+      actualBoundingBoxDescent,
+    } = context.measureText(truncatedText);
+    const textHeight = actualBoundingBoxAscent + actualBoundingBoxDescent;
+
+    const tokens = model.tokensForCell(
+      modelColumn,
+      modelRow,
+      truncatedText.length
+    );
+
+    // Check if the truncated text contains a link
+    if (tokens.length === 0) {
+      context.restore();
+      return (EMPTY_ARRAY as unknown) as TokenBox[];
+    }
+
+    const cachedTokenBoxes = this.getCachedTokenBoxesForVisibleCell(
+      truncatedText,
+      tokens,
+      theme.font,
+      'middle',
+      textHeight,
+      context
+    ).map(tokenBox => ({
+      x1: tokenBox.x1 + textX,
+      y1: tokenBox.y1 + (textY - actualBoundingBoxAscent),
+      x2: tokenBox.x2 + textX,
+      y2: tokenBox.y2 + (textY - actualBoundingBoxAscent),
+      token: tokenBox.token,
+    }));
+
+    context.restore();
+
+    return cachedTokenBoxes;
+  }
+
+  /**
+   * Returns an array of token boxes with the coordinates relative to the top left corner of the text
+   */
+  getCachedTokenBoxesForVisibleCell = memoizeClear(
+    (
+      truncatedText: string,
+      tokens: Token[],
+      // _font and _baseline are passed in so value is re-calculated when they change
+      // They should already be set on the `context`, so they are not used in this method
+      _font: string,
+      _baseline: CanvasTextBaseline,
+      textHeight: number,
+      context: CanvasRenderingContext2D
+    ): TokenBox[] => {
+      const top = 0;
+      const bottom = textHeight;
+
+      const tokenBoxes: TokenBox[] = [];
+
+      // The index where the last token ended
+      let lastTokenEnd = 0;
+      // The width of the text preceding the current token
+      let currentTextWidth = 0;
+      // Loop through array and push them to array
+      for (let i = 0; i < tokens.length; i += 1) {
+        const token = tokens[i];
+        const { start, end } = token;
+        // The last token value is calculated based on the full text so the value needs to be truncated
+        const value =
+          end > truncatedText.length
+            ? truncatedText.substring(start)
+            : token.value;
+
+        // Add the width of the text in between this token and the last token
+        currentTextWidth += context.measureText(
+          truncatedText.substring(lastTokenEnd, start)
+        ).width;
+        const tokenWidth = context.measureText(value).width;
+
+        // Check if the x position is less than the grid x, then tokenWidth should be shifted by gridX - startX
+
+        const left = currentTextWidth;
+        const right = left + tokenWidth;
+
+        const newTokenBox: TokenBox = {
+          x1: left,
+          y1: top,
+          x2: right,
+          y2: bottom,
+          token,
+        };
+
+        tokenBoxes.push(newTokenBox);
+
+        lastTokenEnd = end;
+        currentTextWidth += tokenWidth;
+      }
+
+      return tokenBoxes;
+    },
+    { max: 10000 }
+  );
 }
 
 export default GridRenderer;
