+++ packages/grid/src/GridTheme.ts
@@ -132,12 +132,6 @@ export type GridTheme = {
   // Divider colors between the floating parts and the grid
   floatingDividerOuterColor: GridColor;
   floatingDividerInnerColor: GridColor;
-
-  zeroLineColor: GridColor;
-  positiveBarColor: GridColor;
-  negativeBarColor: GridColor;
-
-  dataBarHorizontalPadding: number;
 };
 
 /**
@@ -227,13 +221,6 @@ const defaultTheme: GridTheme = Object.freeze({
   // Divider colors between the floating parts and the grid
   floatingDividerOuterColor: '#000000',
   floatingDividerInnerColor: '#cccccc',
-
-  // Databar
-  zeroLineColor: '#888888',
-  positiveBarColor: '#00ff00',
-  negativeBarColor: '#ff0000',
-
-  dataBarHorizontalPadding: 90,
 });
 
 export default defaultTheme;
