+++ packages/grid/src/index.ts
@@ -20,14 +20,9 @@ export { default as MockTreeGridModel } from './MockTreeGridModel';
 export { default as memoizeClear } from './memoizeClear';
 export { default as StaticDataGridModel } from './StaticDataGridModel';
 export { default as ViewportDataGridModel } from './ViewportDataGridModel';
-export { default as MockDataBarGridModel } from './MockDataBarGridModel';
 export * from './key-handlers';
 export * from './mouse-handlers';
 export * from './errors';
 export * from './EventHandlerResult';
 export { default as ThemeContext } from './ThemeContext';
-export type { default as CellRenderer, CellRenderType } from './CellRenderer';
-export { default as TextCellRenderer } from './TextCellRenderer';
-export { default as DataBarCellRenderer } from './DataBarCellRenderer';
-export * from './TokenBoxCellRenderer';
 export * from './GridRendererTypes';
