+++ packages/grid/src/mouse-handlers/GridTokenMouseHandler.ts
@@ -1,13 +1,11 @@
 /* eslint class-methods-use-this: "off" */
 
-import { getOrThrow } from '@deephaven/utils';
 import { isEditableGridModel } from '../EditableGridModel';
 import { EventHandlerResult } from '../EventHandlerResult';
 import Grid from '../Grid';
 import GridMouseHandler, { GridMouseEvent } from '../GridMouseHandler';
 import GridRange from '../GridRange';
 import GridUtils, { GridPoint, isLinkToken, TokenBox } from '../GridUtils';
-import { isTokenBoxCellRenderer } from '../TokenBoxCellRenderer';
 
 class GridTokenMouseHandler extends GridMouseHandler {
   timeoutId?: ReturnType<typeof setTimeout>;
@@ -23,21 +21,10 @@ class GridTokenMouseHandler extends GridMouseHandler {
 
   isHoveringLink(gridPoint: GridPoint, grid: Grid): boolean {
     const { column, row, x, y } = gridPoint;
-    const { renderer, metrics, props } = grid;
-    const { model } = props;
+    const { renderer, metrics } = grid;
 
     if (column == null || row == null || metrics == null) {
       this.currentLinkBox = undefined;
-      return false;
-    }
-
-    const { modelRows, modelColumns } = metrics;
-    const modelRow = getOrThrow(modelRows, row);
-    const modelColumn = getOrThrow(modelColumns, column);
-
-    const renderType = model.renderTypeForCell(modelColumn, modelRow);
-    const cellRenderer = renderer.getCellRenderer(renderType);
-    if (!isTokenBoxCellRenderer(cellRenderer)) {
       return false;
     }
 
@@ -49,8 +36,7 @@ class GridTokenMouseHandler extends GridMouseHandler {
     }
 
     const renderState = grid.updateRenderState();
-
-    const tokensInCell = cellRenderer.getTokenBoxesForVisibleCell(
+    const tokensInCell = renderer.getTokenBoxesForVisibleCell(
       column,
       row,
       renderState
