+++ packages/iris-grid/src/IrisGridTheme.module.scss
@@ -91,8 +91,4 @@ $header-height: 30px;
 
   overflow-button-color: $gray-300;
   overflow-button-hover-color: $gray-100;
-
-  zero-line-color: $gray-500;
-  positive-bar-color: $green;
-  negative-bar-color: $red;
 }
