+++ packages/iris-grid/src/IrisGridTheme.ts
@@ -135,10 +135,6 @@ const theme: Partial<IrisGridThemeType> = Object.freeze({
 
   overflowButtonColor: IrisGridTheme['overflow-button-color'],
   overflowButtonHoverColor: IrisGridTheme['overflow-button-hover-color'],
-
-  zeroLineColor: IrisGridTheme['zero-line-color'],
-  positiveBarColor: IrisGridTheme['positive-bar-color'],
-  negativeBarColor: IrisGridTheme['negative-bar-color'],
 });
 
 export default theme;
