+++ packages/iris-grid/src/mousehandlers/IrisGridTokenMouseHandler.ts
@@ -1,13 +1,11 @@
 import {
   EventHandlerResult,
-  getOrThrow,
   Grid,
   GridMouseHandler,
   GridPoint,
   GridUtils,
   isLinkToken,
   TokenBox,
-  isTokenBoxCellRenderer,
 } from '@deephaven/grid';
 import deepEqual from 'deep-equal';
 import IrisGrid from '../IrisGrid';
@@ -30,21 +28,10 @@ class IrisGridTokenMouseHandler extends GridMouseHandler {
 
   isHoveringLink(gridPoint: GridPoint, grid: Grid): boolean {
     const { column, row, x, y } = gridPoint;
-    const { renderer, metrics, props } = grid;
-    const { model } = props;
+    const { renderer, metrics } = grid;
 
     if (column == null || row == null || metrics == null) {
       this.currentLinkBox = undefined;
-      return false;
-    }
-
-    const { modelRows, modelColumns } = metrics;
-    const modelRow = getOrThrow(modelRows, row);
-    const modelColumn = getOrThrow(modelColumns, column);
-
-    const renderType = model.renderTypeForCell(modelColumn, modelRow);
-    const cellRenderer = renderer.getCellRenderer(renderType);
-    if (!isTokenBoxCellRenderer(cellRenderer)) {
       return false;
     }
 
@@ -56,7 +43,7 @@ class IrisGridTokenMouseHandler extends GridMouseHandler {
     }
 
     const renderState = grid.updateRenderState();
-    const tokensInCell = cellRenderer.getTokenBoxesForVisibleCell(
+    const tokensInCell = renderer.getTokenBoxesForVisibleCell(
       column,
       row,
       renderState
